Adding network mask parsing
This is only parsing of the network mask. It is accessed easily by a getter, no
methods are provided to do any math or to get the network prefix.

Zero compression can create problems if not taken into account when
dealing with masks, therefore it is handled specially.

Unit tests check for both the non compressed and the compressed cases.